% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimates.R
\name{get_estimates}
\alias{get_estimates}
\title{get_estimates}
\usage{
get_estimates(model)
}
\arguments{
\item{model}{An lmerMod object}
}
\value{
A list object with fixed and random effect estimates
}
\description{
This function allows us to extract the estimated parameter estimates from a lmer model.
}
