% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcomp.R
\name{get_vcomp}
\alias{get_vcomp}
\title{get_vcomp}
\usage{
get_vcomp(model)
}
\value{
A data.frame with the variance component estimates and the percentage of total variance.
}
\description{
This function allows you to extract the estimated random effects from a lmer model object
}
